Add ready() method
This patch moves the checksum calculation and the package id sanity
checks into the interrupt handler, which allows the receiver to drop
invalid and duplicate packages much earlier. So if the inputBufReady
flag is true, a valid deduplicated package is in the buffer.

The ready() method makes the inputBufReady flag externally
accessible and therefore allows for asynchronous operation. If ready()
returns true, then recvPackage() will not block.